Fix statistics error of jank frames in CTS tests
During  Davey DrawFrame CTS tests, we found that occasionally the new
frame would be posted to UI thread before a pending Vsync callback (
which is postponed by the 700/150 ms frame generated in Davey Test).
Normally, such pending Vsync  callback should run before the new coming
frame and do nothing, then the subsequent new frame would trigger a new
Vsync callback to draw itself .

While on the low end device (especially 2-core devices) , the pending
Vsync callback may not be scheduled in time and be executed after the
new incoming frame, thus  the new frame would been drawn just in such
pending Vsync callback rather than schedule a new Vsync callback.
However, the "intended Vsync” time of such pending Vsync was hundreds
of milliseconds (150/700 ms) ago which will make the frames
“finish time - intended Vsync time” much larger than expected. Thus such
new frame will be regarded as a “jank” frame  although the frame was
drawn as soon as the app posted it out.

We don't think it is reasonable to regard such kind of frames as
"jank" frames, so we compare the "intended Vsync" time and the frame
posting time in this patch. If the frame posting time is after a pending
Vsync, we should use posting time as the real start time of such frame.

Bug: 110967410
Test: run cts --m CtsIncidentHostTestCases -t com.android.server.cts.GraphicsStatsValidationTest#testDaveyDrawFrame
Change-Id: If7217504e6c56152d6871479eff88e6ae8a1bc2c

Signed-off-by: Guangli Li <guangli.li@intel.com>
Signed-off-by: Zhongmin Wu <zhongmin.wu@intel.com>
Sigend-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>