base: Set a dark tint when disable wallpaper tint is enabled
Dirty Unicorns made a nice feature to toggle the use of the wallpaper tint over various parts of
the system (like Recents, keyguard and power menu) but removes the overlay alpha of the tinted
sections. That may be a bit ugly on Power menu or even if you use custom lockscreen alpha this
doesn't work.

Luckily, Google left some test documents that make you understand how their new API works and
we're using one on this case to generate a custom WallpaperColors with all black colors to rely
on that in case that tint is disabled. And that allows us to keep the normal dark overlay and
make it work like it should be on previous android versions.

Credits and thanks goes to @Mazda-- for making the main code for disabling the tint and his
feedback on improvements, and @ishubhamsingh for some code ideas. Exposed every tint color just
for the sake of see if themers really want to tweak this